refactor: split docker helpers and share compose-args/logRules between setup and run
