+++ core/lib/docker/client.ts
@@ -1,7 +1,7 @@
 import { execFileSync, spawn, type ChildProcess } from "node:child_process";
 import { once } from "node:events";
 import { createInterface } from "node:readline";
-import { buildDockerCpArgs } from "./container.ts";
+import { buildDockerCpArgs } from "./cp-args.ts";
 import { parseDockerInspectEnv } from "./container-env.ts";
 
 /** `docker ps --format '{{.ID}}'` prints one ID per line, possibly with
