+++ core/lib/docker/compose-project-name.ts
@@ -23,17 +23,3 @@ export function resolveProjectName(
 ): string {
   return composeProjectNameOverride || deriveProjectName(builderName);
 }
-
-export interface BuildDockerCpArgsOptions {
-  containerName: string;
-  containerPath: string;
-  hostPath: string;
-}
-
-export function buildDockerCpArgs({
-  containerName,
-  containerPath,
-  hostPath,
-}: BuildDockerCpArgsOptions): string[] {
-  return ["cp", `${containerName}:${containerPath}`, hostPath];
-}
